requirements.txt
langchain
langchain-community
langchain-groq
unstructured
selenium
chromadb
streamlit
pandas
python-dotenv
fastapi
uvicorn